Hack - decrement version number
Apache beam insist on using 0.5.31 and not 0.5.32, so we can't just
install from github with pinned commit, to hackily get around this, we
decrement the version number AND install from github archive.

Yes, we do need the changes made that are not in the actual 0.5.31
version, namely the unpickling changes, as Python3.7+ has optimised how
unpickling works for list (and its subclasses) and inadvertently broke
apitools.